refactor(maintaining plots)
