+++ bioptim/gui/online_callback.py
@@ -1,0 +1,229 @@
+import multiprocessing as mp
+
+from casadi import Callback, nlpsol_out, nlpsol_n_out, Sparsity
+from matplotlib import pyplot as plt
+
+from .plot import PlotOcp
+
+
+class OnlineCallback(Callback):
+    """
+    CasADi interface of Ipopt callbacks
+
+    Attributes
+    ----------
+    ocp: OptimalControlProgram
+        A reference to the ocp to show
+    nx: int
+        The number of optimization variables
+    ng: int
+        The number of constraints
+    queue: mp.Queue
+        The multiprocessing queue
+    plotter: ProcessPlotter
+        The callback for plotting for the multiprocessing
+    plot_process: mp.Process
+        The multiprocessing placeholder
+
+    Methods
+    -------
+    get_n_in() -> int
+        Get the number of variables in
+    get_n_out() -> int
+        Get the number of variables out
+    get_name_in(i: int) -> int
+        Get the name of a variable
+    get_name_out(_) -> str
+        Get the name of the output variable
+    get_sparsity_in(self, i: int) -> tuple[int]
+        Get the sparsity of a specific variable
+    eval(self, arg: list | tuple) -> list[int]
+        Send the current data to the plotter
+    """
+
+    def __init__(self, ocp, opts: dict = None, show_options: dict = None):
+        """
+        Parameters
+        ----------
+        ocp: OptimalControlProgram
+            A reference to the ocp to show
+        opts: dict
+            Option to AnimateCallback method of CasADi
+        show_options: dict
+            The options to pass to PlotOcp
+        """
+        if opts is None:
+            opts = {}
+
+        Callback.__init__(self)
+        self.ocp = ocp
+        self.nx = self.ocp.variables_vector.shape[0]
+        self.ng = 0
+        self.construct("AnimateCallback", opts)
+
+        self.queue = mp.Queue()
+        self.plotter = self.ProcessPlotter(self.ocp)
+        self.plot_process = mp.Process(target=self.plotter, args=(self.queue, show_options), daemon=True)
+        self.plot_process.start()
+
+    def close(self):
+        self.plot_process.kill()
+
+    @staticmethod
+    def get_n_in() -> int:
+        """
+        Get the number of variables in
+
+        Returns
+        -------
+        The number of variables in
+        """
+
+        return nlpsol_n_out()
+
+    @staticmethod
+    def get_n_out() -> int:
+        """
+        Get the number of variables out
+
+        Returns
+        -------
+        The number of variables out
+        """
+
+        return 1
+
+    @staticmethod
+    def get_name_in(i: int) -> int:
+        """
+        Get the name of a variable
+
+        Parameters
+        ----------
+        i: int
+            The index of the variable
+
+        Returns
+        -------
+        The name of the variable
+        """
+
+        return nlpsol_out(i)
+
+    @staticmethod
+    def get_name_out(_) -> str:
+        """
+        Get the name of the output variable
+
+        Returns
+        -------
+        The name of the output variable
+        """
+
+        return "ret"
+
+    def get_sparsity_in(self, i: int) -> tuple:
+        """
+        Get the sparsity of a specific variable
+
+        Parameters
+        ----------
+        i: int
+            The index of the variable
+
+        Returns
+        -------
+        The sparsity of the variable
+        """
+
+        n = nlpsol_out(i)
+        if n == "f":
+            return Sparsity.scalar()
+        elif n in ("x", "lam_x"):
+            return Sparsity.dense(self.nx)
+        elif n in ("g", "lam_g"):
+            return Sparsity.dense(self.ng)
+        else:
+            return Sparsity(0, 0)
+
+    def eval(self, arg: list | tuple) -> list:
+        """
+        Send the current data to the plotter
+
+        Parameters
+        ----------
+        arg: list | tuple
+            The data to send
+
+        Returns
+        -------
+        A list of error index
+        """
+        send = self.queue.put
+        send(arg[0])
+        return [0]
+
+    class ProcessPlotter(object):
+        """
+        The plotter that interface PlotOcp and the multiprocessing
+
+        Attributes
+        ----------
+        ocp: OptimalControlProgram
+            A reference to the ocp to show
+        pipe: mp.Queue
+            The multiprocessing queue to evaluate
+        plot: PlotOcp
+            The handler on all the figures
+
+        Methods
+        -------
+        callback(self) -> bool
+            The callback to update the graphs
+        """
+
+        def __init__(self, ocp):
+            """
+            Parameters
+            ----------
+            ocp: OptimalControlProgram
+                A reference to the ocp to show
+            """
+
+            self.ocp = ocp
+
+        def __call__(self, pipe: mp.Queue, show_options: dict):
+            """
+            Parameters
+            ----------
+            pipe: mp.Queue
+                The multiprocessing queue to evaluate
+            show_options: dict
+                The option to pass to PlotOcp
+            """
+
+            if show_options is None:
+                show_options = {}
+            self.pipe = pipe
+            self.plot = PlotOcp(self.ocp, **show_options)
+            timer = self.plot.all_figures[0].canvas.new_timer(interval=10)
+            timer.add_callback(self.callback)
+            timer.start()
+            plt.show()
+
+        def callback(self) -> bool:
+            """
+            The callback to update the graphs
+
+            Returns
+            -------
+            True if everything went well
+            """
+
+            while not self.pipe.empty():
+                v = self.pipe.get()
+                self.plot.update_data(v)
+
+            for i, fig in enumerate(self.plot.all_figures):
+                fig.canvas.draw()
+            return True
