+++ bioptim/gui/plot.py
@@ -1,21 +1,34 @@
-from typing import Callable, Any
-import multiprocessing as mp
 import tkinter
 from itertools import accumulate
+from typing import Callable, Any
 
 import numpy as np
+from casadi import DM
 from matplotlib import pyplot as plt, lines
 from matplotlib.ticker import StrMethodFormatter
-from casadi import Callback, nlpsol_out, nlpsol_n_out, Sparsity, DM
 
+from ..dynamics.ode_solver import OdeSolver
 from ..limits.path_conditions import Bounds
 from ..limits.penalty_helpers import PenaltyHelpers
 from ..misc.enums import PlotType, Shooting, SolutionIntegrator, QuadratureRule, InterpolationType
 from ..misc.mapping import Mapping, BiMapping
-from ..optimization.solution.solution import Solution
-from ..dynamics.ode_solver import OdeSolver
 from ..optimization.optimization_vector import OptimizationVectorHelper
+from ..optimization.solution.solution import Solution
 from ..optimization.solution.solution_data import SolutionMerge
+
+DEFAULT_COLORS = {
+    PlotType.PLOT: "tab:green",
+    PlotType.INTEGRATED: "tab:brown",
+    PlotType.STEP: "tab:orange",
+    PlotType.POINT: "tab:purple"
+}
+
+DEFAULT_LINESTYLES = {
+    PlotType.PLOT: "-",
+    PlotType.INTEGRATED: None,
+    PlotType.STEP: "-",
+    PlotType.POINT: None
+}
 
 
 class CustomPlot:
@@ -49,21 +62,21 @@ class CustomPlot:
     """
 
     def __init__(
-        self,
-        update_function: Callable,
-        plot_type: PlotType = PlotType.PLOT,
-        axes_idx: BiMapping | tuple | list = None,
-        legend: tuple | list = None,
-        combine_to: str = None,
-        color: str = None,
-        linestyle: str = None,
-        ylim: tuple | list = None,
-        bounds: Bounds = None,
-        node_idx: list | slice | range = None,
-        label: list = None,
-        compute_derivative: bool = False,
-        integration_rule: QuadratureRule = QuadratureRule.RECTANGLE_LEFT,
-        **parameters: Any,
+            self,
+            update_function: Callable,
+            plot_type: PlotType = PlotType.PLOT,
+            axes_idx: BiMapping | tuple | list = None,
+            legend: tuple | list = None,
+            combine_to: str = None,
+            color: str = None,
+            linestyle: str = None,
+            ylim: tuple | list = None,
+            bounds: Bounds = None,
+            node_idx: list | slice | range = None,
+            label: list = None,
+            compute_derivative: bool = False,
+            integration_rule: QuadratureRule = QuadratureRule.RECTANGLE_LEFT,
+            **parameters: Any,
     ):
         """
         Parameters
@@ -193,12 +206,12 @@ class PlotOcp:
     """
 
     def __init__(
-        self,
-        ocp,
-        automatically_organize: bool = True,
-        show_bounds: bool = False,
-        shooting_type: Shooting = Shooting.MULTIPLE,
-        integrator: SolutionIntegrator = SolutionIntegrator.OCP,
+            self,
+            ocp,
+            automatically_organize: bool = True,
+            show_bounds: bool = False,
+            shooting_type: Shooting = Shooting.MULTIPLE,
+            integrator: SolutionIntegrator = SolutionIntegrator.OCP,
     ):
         """
         Prepares the figures during the simulation
@@ -457,13 +470,11 @@ class PlotOcp:
                     else:
                         label = None
 
+                    color = (self.custom_plots[variable][i].color if self.custom_plots[variable][i].color else
+                             DEFAULT_COLORS[plot_type])
+
                     if plot_type == PlotType.PLOT:
                         zero = np.zeros((t.shape[0], 1))
-                        color = (
-                            self.custom_plots[variable][i].color
-                            if self.custom_plots[variable][i].color
-                            else "tab:green"
-                        )
                         self.plots.append(
                             [
                                 plot_type,
@@ -480,11 +491,6 @@ class PlotOcp:
                         )
                     elif plot_type == PlotType.INTEGRATED:
                         plots_integrated = []
-                        color = (
-                            self.custom_plots[variable][i].color
-                            if self.custom_plots[variable][i].color
-                            else "tab:brown"
-                        )
                         for cmp in range(nlp.ns):
                             plots_integrated.append(
                                 ax.plot(
@@ -499,11 +505,6 @@ class PlotOcp:
 
                     elif plot_type == PlotType.STEP:
                         zero = np.zeros((t.shape[0], 1))
-                        color = (
-                            self.custom_plots[variable][i].color
-                            if self.custom_plots[variable][i].color
-                            else "tab:orange"
-                        )
                         linestyle = (
                             self.custom_plots[variable][i].linestyle
                             if self.custom_plots[variable][i].linestyle
@@ -518,11 +519,6 @@ class PlotOcp:
                         )
                     elif plot_type == PlotType.POINT:
                         zero = np.zeros((t.shape[0], 1))
-                        color = (
-                            self.custom_plots[variable][i].color
-                            if self.custom_plots[variable][i].color
-                            else "tab:purple"
-                        )
                         self.plots.append(
                             [
                                 plot_type,
@@ -714,7 +710,7 @@ class PlotOcp:
         self.__update_axes()
 
     def _compute_y_from_plot_func(
-        self, custom_plot: CustomPlot, phase_idx, time_stepwise, dt, x_decision, x_stepwise, u, p, a
+            self, custom_plot: CustomPlot, phase_idx, time_stepwise, dt, x_decision, x_stepwise, u, p, a
     ):
         """
         Compute the y data from the plot function
@@ -949,226 +945,3 @@ class PlotOcp:
 
         n_rows = int(round(np.sqrt(nb)))
         return n_rows + 1 if n_rows * n_rows < nb else n_rows, n_rows
-
-
-class OnlineCallback(Callback):
-    """
-    CasADi interface of Ipopt callbacks
-
-    Attributes
-    ----------
-    ocp: OptimalControlProgram
-        A reference to the ocp to show
-    nx: int
-        The number of optimization variables
-    ng: int
-        The number of constraints
-    queue: mp.Queue
-        The multiprocessing queue
-    plotter: ProcessPlotter
-        The callback for plotting for the multiprocessing
-    plot_process: mp.Process
-        The multiprocessing placeholder
-
-    Methods
-    -------
-    get_n_in() -> int
-        Get the number of variables in
-    get_n_out() -> int
-        Get the number of variables out
-    get_name_in(i: int) -> int
-        Get the name of a variable
-    get_name_out(_) -> str
-        Get the name of the output variable
-    get_sparsity_in(self, i: int) -> tuple[int]
-        Get the sparsity of a specific variable
-    eval(self, arg: list | tuple) -> list[int]
-        Send the current data to the plotter
-    """
-
-    def __init__(self, ocp, opts: dict = None, show_options: dict = None):
-        """
-        Parameters
-        ----------
-        ocp: OptimalControlProgram
-            A reference to the ocp to show
-        opts: dict
-            Option to AnimateCallback method of CasADi
-        show_options: dict
-            The options to pass to PlotOcp
-        """
-        if opts is None:
-            opts = {}
-
-        Callback.__init__(self)
-        self.ocp = ocp
-        self.nx = self.ocp.variables_vector.shape[0]
-        self.ng = 0
-        self.construct("AnimateCallback", opts)
-
-        self.queue = mp.Queue()
-        self.plotter = self.ProcessPlotter(self.ocp)
-        self.plot_process = mp.Process(target=self.plotter, args=(self.queue, show_options), daemon=True)
-        self.plot_process.start()
-
-    def close(self):
-        self.plot_process.kill()
-
-    @staticmethod
-    def get_n_in() -> int:
-        """
-        Get the number of variables in
-
-        Returns
-        -------
-        The number of variables in
-        """
-
-        return nlpsol_n_out()
-
-    @staticmethod
-    def get_n_out() -> int:
-        """
-        Get the number of variables out
-
-        Returns
-        -------
-        The number of variables out
-        """
-
-        return 1
-
-    @staticmethod
-    def get_name_in(i: int) -> int:
-        """
-        Get the name of a variable
-
-        Parameters
-        ----------
-        i: int
-            The index of the variable
-
-        Returns
-        -------
-        The name of the variable
-        """
-
-        return nlpsol_out(i)
-
-    @staticmethod
-    def get_name_out(_) -> str:
-        """
-        Get the name of the output variable
-
-        Returns
-        -------
-        The name of the output variable
-        """
-
-        return "ret"
-
-    def get_sparsity_in(self, i: int) -> tuple:
-        """
-        Get the sparsity of a specific variable
-
-        Parameters
-        ----------
-        i: int
-            The index of the variable
-
-        Returns
-        -------
-        The sparsity of the variable
-        """
-
-        n = nlpsol_out(i)
-        if n == "f":
-            return Sparsity.scalar()
-        elif n in ("x", "lam_x"):
-            return Sparsity.dense(self.nx)
-        elif n in ("g", "lam_g"):
-            return Sparsity.dense(self.ng)
-        else:
-            return Sparsity(0, 0)
-
-    def eval(self, arg: list | tuple) -> list:
-        """
-        Send the current data to the plotter
-
-        Parameters
-        ----------
-        arg: list | tuple
-            The data to send
-
-        Returns
-        -------
-        A list of error index
-        """
-        send = self.queue.put
-        send(arg[0])
-        return [0]
-
-    class ProcessPlotter(object):
-        """
-        The plotter that interface PlotOcp and the multiprocessing
-
-        Attributes
-        ----------
-        ocp: OptimalControlProgram
-            A reference to the ocp to show
-        pipe: mp.Queue
-            The multiprocessing queue to evaluate
-        plot: PlotOcp
-            The handler on all the figures
-
-        Methods
-        -------
-        callback(self) -> bool
-            The callback to update the graphs
-        """
-
-        def __init__(self, ocp):
-            """
-            Parameters
-            ----------
-            ocp: OptimalControlProgram
-                A reference to the ocp to show
-            """
-
-            self.ocp = ocp
-
-        def __call__(self, pipe: mp.Queue, show_options: dict):
-            """
-            Parameters
-            ----------
-            pipe: mp.Queue
-                The multiprocessing queue to evaluate
-            show_options: dict
-                The option to pass to PlotOcp
-            """
-
-            if show_options is None:
-                show_options = {}
-            self.pipe = pipe
-            self.plot = PlotOcp(self.ocp, **show_options)
-            timer = self.plot.all_figures[0].canvas.new_timer(interval=10)
-            timer.add_callback(self.callback)
-            timer.start()
-            plt.show()
-
-        def callback(self) -> bool:
-            """
-            The callback to update the graphs
-
-            Returns
-            -------
-            True if everything went well
-            """
-
-            while not self.pipe.empty():
-                v = self.pipe.get()
-                self.plot.update_data(v)
-
-            for i, fig in enumerate(self.plot.all_figures):
-                fig.canvas.draw()
-            return True
